[SYCL][NFC] Move Regression/vec_logical_ops E2E test
The test is way too simple to require execution on an actual device.
Therefore, moved it to regular LIT tests and restricted it to
`-fsyntax-only` mode.

Moreover, the test lacked checks for `fp64` aspect which caused failures
on devices that don't support it.
